+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased][unreleased]
 
+- Make TODOs repo-wide, not per-file todos
+- Improve TODO screen and files screen TUI/TUX including hotkeys
+- Fix TODOs autosave bug with todo duplication
+
 ## [0.1.6][] - 2026-09-15
 
 - Auto updates for patch/minor, y/n per new major